chore(app): remove sync data popup
