+++ packages/browseros-agent/apps/app/lib/onboarding/onboardingStorage.ts
@@ -5,11 +5,6 @@ export const importHintDismissedAtStorage = storage.defineItem<number | null>(
   { fallback: null },
 )
 
-export const signInHintDismissedAtStorage = storage.defineItem<number | null>(
-  'local:signInHintDismissedAt',
-  { fallback: null },
-)
-
 export const authRedirectPathStorage = storage.defineItem<string | null>(
   'local:authRedirectPath',
   { fallback: null },
